Integrate all backwards compatibility methods into the TrilogyAdapter class
